Migrate to SectionProperty to separate file
